Search header logic refactor 2
