+++ src/Search/SearchHeader.tsx
@@ -6,18 +6,6 @@ import {Icon} from '../components'
 import {SearchBar} from './SearchBar'
 import {useSearch} from './SearchContext'
 
-export const useSearchResult = ({target, searchBy, sort = true}) => {
-  const {search} = useSearch()
-  const searchTermLowerCase = search.toLowerCase()
-  const filteredTarget =
-    searchTermLowerCase.length > 0 && target.length > 0
-      ? target.filter((n) => n[searchBy].toLowerCase().includes(searchTermLowerCase))
-      : target
-  const searchResult = sort ? filteredTarget.sort((a, b) => a[searchBy].localeCompare(b[searchBy])) : filteredTarget
-
-  return {searchResult}
-}
-
 export const useSearchHeaderOptions = ({placeHolderText, title}) => {
   const [searchVisible, setSearchVisible] = React.useState(false)
   const {clearSearch} = useSearch()
